break_utility/requirements.txt
absl-py==0.8.1
alabaster==0.7.12
allennlp==0.9.0
astor==0.8.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
boto3==1.10.23
botocore==1.13.23
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
conllu==1.3.1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
editdistance==0.5.3
en-core-web-sm==2.1.0
entrypoints==0.3
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
gast==0.2.2
gevent==1.4.0
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.16.1
h5py==2.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
json5==0.8.5
jsonlines==1.2.0
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.2
numpy==1.17.4
numpydoc==0.9.1
olefile==0.46
opt-einsum==3.1.0
overrides==2.5
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plac==0.9.6
pluggy==0.13.0
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.0
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
pyzmq==18.1.0
regex==2019.11.1
requests==2.22.0
responses==0.10.6
s3transfer==0.2.1
sacremoses==0.0.38
scikit-learn==0.21.3
scipy==1.3.2
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.13.0
snowballstemmer==2.0.0
spacy==2.1.9
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.2.0
tensorboard==2.0.0
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.0.8
torch==1.2.0
torchvision==0.4.0a0+6b959ee
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
transformers==2.3.0
ujson==1.35
Unidecode==1.1.1
urllib3==1.25.7
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
word2number==1.1
wrapt==1.11.2
zipp==0.6.0
